dismemberment complete
